+++ openfisca_us/variables/irs/credits/ctc/credit.py
@@ -5,7 +5,7 @@ class ctc(Variable):
     value_type = float
     entity = TaxUnit
     label = "Child Tax Credit"
-    unit = "currency-USD"
+    unit = USD
     documentation = "Total value of the non-refundable and refundable portions of the Child Tax Credit."
     definition_period = YEAR
     reference = "https://www.law.cornell.edu/uscode/text/26/24#a"