+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.45.1] - 2022-04-15 14:23:11
+
+### Added
+
+- Legislative references for CDCC parameters.
+
+### Fixed
+
+- CDCC uses maximum dependent parameter.
+
 ## [0.45.0] - 2022-04-14 08:19:40
 
 ### Added
@@ -628,6 +638,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
+[0.45.1]: https://github.com/PolicyEngine/openfisca-us/compare/0.45.0...0.45.1
 [0.45.0]: https://github.com/PolicyEngine/openfisca-us/compare/0.44.0...0.45.0
 [0.44.0]: https://github.com/PolicyEngine/openfisca-us/compare/0.43.1...0.44.0
 [0.43.1]: https://github.com/PolicyEngine/openfisca-us/compare/0.43.0...0.43.1